+++ mycodo/inputs/examples/minimal_humidity_temperature.py
@@ -30,6 +30,8 @@ INPUT_INFORMATION = {
     'measurements_dict': measurements_dict,
     'measurements_use_same_timestamp': True,
 
+    'message': "Note: This is just an informative note to the user.",
+
     'options_enabled': [
         'measurements_select',
         'period',